Completed Tutorial: Building a Simple Component
Comfort: 5
Clarity: 4

It was pretty interesting seeing how Ember uses handlebars to create
components for each real estate listing. I am starting to get used to
using Javascript and html in the same file. It was weird at first
because I was so used to keeping Javascript and HTML separate. Plus,
I see the connect of clicking the image or link makes the image larger,
but I don't get why a click action triggers it.
 Maybe it is the nature of anchor tags. I thought we would have to use
 event listeners.