Completed Service and Utilities
Comfort: 3
Clarity: 4

I don't like Ember. I like that Ember has structure to it, but it is too
restrictive. For 3rd Party API's, it seems to do the work for you but
you have to set up files and code in certain ways. Plus there are too
many files to keep track off when you want to create a feature.
I prefer to have more flexibility in developing the front end.
In the back end, especially for SQL its great to have something like
Ember ... a hem Rails because SQL data are straightforward and
there is only one thing to do things.